chore(package): Add prePublish hook
This is for developers only. It checks if there are any untracked or changed files in the local repository and if the main entry point exists. For example, if you forgot to run the build task, there will be no lib folder, so the main entry point would not be accessible...